porcupine/requirements.txt
python-dotenv
pvporcupine
pyaudio
